requirements.txt
Django==4.2.1
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
django-environ==0.10.0
drf-yasg==1.21.5
psycopg-binary==3.1.9
